refactor: alterando diretório das funções de máscara
